+++ lib/core/localization/app_localizations.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:intl/intl.dart';
 
 import 'translations/en_us.dart';
 import 'translations/ne_np.dart';
