+++ lib/feature/home/presentation/screen/home_screen.dart
@@ -3,7 +3,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:fun_math/feature/home/presentation/widgets/difficulty_selector.dart';
 import 'package:fun_math/core/theme/provider/theme_provider.dart';
+import 'package:fun_math/feature/home/presentation/widgets/math_character.dart';
 import 'package:iconsax/iconsax.dart';
+import '../widgets/footer.dart';
 import '../widgets/game_card.dart';
 import '../widgets/learning_section.dart';
 import '../widgets/animated_header.dart';
@@ -100,10 +102,8 @@ class HomeScreen extends ConsumerWidget {
 
                     // Learning Section with fun illustrations
                    
-                 
-
-                    // Footer with animated math symbols and fun message
-                    _buildFooter(context, isDarkMode),
+                
+                    buildFooter(context, isDarkMode),
                   ],
                 ),
               ),
@@ -114,7 +114,7 @@ class HomeScreen extends ConsumerWidget {
     );
   }
   Widget _buildAppBar(BuildContext context, bool isDarkMode, ThemeProvider themeNotifier, ColorScheme colorScheme) {
-    return AnimatedHeader(userName: "Alex").animate().fadeIn(duration: 600.ms).slideY(begin: -0.2, end: 0, duration: 700.ms, curve: Curves.easeOutQuad);
+    return AnimatedHeader().animate().fadeIn(duration: 600.ms).slideY(begin: -0.2, end: 0, duration: 700.ms, curve: Curves.easeOutQuad);
   }
 
   Widget _buildWelcomeSection(BuildContext context, bool isDarkMode) {
@@ -165,12 +165,12 @@ class HomeScreen extends ConsumerWidget {
         gradient: LinearGradient(
           colors: isDarkMode
               ? [
-                  const Color(0xFF6A5AE0).withOpacity(0.3),
-                  const Color(0xFFFF8FA2).withOpacity(0.2),
+                  const Color(0xFF6A5AE0).withValues(alpha: .3),
+                  const Color(0xFFFF8FA2).withValues(alpha: 0.2),
                 ]
               : [
-                  const Color(0xFF6A5AE0).withOpacity(0.1),
-                  const Color(0xFFFF8FA2).withOpacity(0.05),
+                  const Color(0xFF6A5AE0).withValues(alpha:0.1),
+                  const Color(0xFFFF8FA2).withValues(alpha:0.05),
                 ],
           begin: Alignment.topLeft,
           end: Alignment.bottomRight,
@@ -178,8 +178,8 @@ class HomeScreen extends ConsumerWidget {
         borderRadius: BorderRadius.circular(16),
         border: Border.all(
           color: isDarkMode
-              ? const Color(0xFF6A5AE0).withOpacity(0.5)
-              : const Color(0xFF6A5AE0).withOpacity(0.2),
+              ? const Color(0xFF6A5AE0).withValues(alpha: .5)
+              : const Color(0xFF6A5AE0).withValues(alpha: .2),
           width: 1,
         ),
       ),
@@ -246,7 +246,7 @@ class HomeScreen extends ConsumerWidget {
             borderRadius: BorderRadius.circular(16),
             boxShadow: [
               BoxShadow(
-                color: Colors.black.withOpacity(0.1),
+                color: Colors.black.withValues(alpha: .1),
                 blurRadius: 8,
                 offset: const Offset(0, 4),
               ),
@@ -261,90 +261,7 @@ class HomeScreen extends ConsumerWidget {
           ),
         ),
         // Math character (simple illustration)
-        Container(
-          width: 100,
-          height: 100,
-          decoration: BoxDecoration(
-            color: const Color(0xFFFFD56F),
-            shape: BoxShape.circle,
-            border: Border.all(
-              color: const Color(0xFFFFB347),
-              width: 4,
-            ),
-          ),
-          child: Stack(
-            alignment: Alignment.center,
-            children: [
-              // Face
-              Positioned(
-                top: 25,
-                child: Container(
-                  width: 40,
-                  height: 40,
-                  decoration: BoxDecoration(
-                    color: Colors.white,
-                    shape: BoxShape.circle,
-                  ),
-                ),
-              ),
-              // Eyes
-              Positioned(
-                top: 35,
-                left: 30,
-                child: Container(
-                  width: 10,
-                  height: 10,
-                  decoration: const BoxDecoration(
-                    color: Colors.black,
-                    shape: BoxShape.circle,
-                  ),
-                ),
-              ),
-              Positioned(
-                top: 35,
-                right: 30,
-                child: Container(
-                  width: 10,
-                  height: 10,
-                  decoration: const BoxDecoration(
-                    color: Colors.black,
-                    shape: BoxShape.circle,
-                  ),
-                ),
-              ),
-              // Smile
-              Positioned(
-                top: 50,
-                child: Container(
-                  width: 30,
-                  height: 15,
-                  decoration: BoxDecoration(
-                    border: Border.all(
-                      color: Colors.black,
-                      width: 2,
-                    ),
-                    borderRadius: const BorderRadius.only(
-                      bottomLeft: Radius.circular(20),
-                      bottomRight: Radius.circular(20),
-                    ),
-                  ),
-                ),
-              ),
-              // Math symbol on body
-              const Positioned(
-                bottom: 20,
-                child: Text(
-                  '+',
-                  style: TextStyle(
-                    fontSize: 30,
-                    fontWeight: FontWeight.bold,
-                    color: Colors.black,
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ),
+        mathCharacter()
       ],
     );
   }
@@ -383,60 +300,7 @@ class HomeScreen extends ConsumerWidget {
     );
   }
 
-  Widget _buildFooter(BuildContext context, bool isDarkMode) {
-    return Column(
-      children: [
-        Divider(
-          color: isDarkMode ? Colors.white24 : Colors.black12,
-          thickness: 1,
-        ),
-        const SizedBox(height: 16),
-        Text(
-          'Math is everywhere! Keep exploring!',
-          style: Theme.of(context).textTheme.titleMedium?.copyWith(
-                fontStyle: FontStyle.italic,
-                color: isDarkMode ? Colors.white70 : Colors.black54,
-              ),
-        ),
-        const SizedBox(height: 16),
-        // Animated math symbols with more fun
-        SizedBox(
-          height: 60,
-          child: ListView(
-            scrollDirection: Axis.horizontal,
-            shrinkWrap: true,
-            children: const [
-              _BouncingMathSymbol(symbol: '+', color: Color(0xFFFF8FA2)),
-              _BouncingMathSymbol(symbol: '-', color: Color(0xFF6A5AE0), delay: 100),
-              _BouncingMathSymbol(symbol: '×', color: Color(0xFFFFD56F), delay: 200),
-              _BouncingMathSymbol(symbol: '÷', color: Color(0xFF92E3A9), delay: 300),
-              _BouncingMathSymbol(symbol: '=', color: Color(0xFF8B80FF), delay: 400),
-              _BouncingMathSymbol(symbol: 'π', color: Color(0xFFFFB347), delay: 500),
-              _BouncingMathSymbol(symbol: '√', color: Color(0xFF6A5AE0), delay: 600),
-              _BouncingMathSymbol(symbol: '∞', color: Color(0xFFFF8FA2), delay: 700),
-              _BouncingMathSymbol(symbol: 'θ', color: Color(0xFF92E3A9), delay: 800),
-            ],
-          ),
-        ),
-        const SizedBox(height: 24),
-        // Fun encouragement message
-        Container(
-          padding: const EdgeInsets.symmetric(vertical: 12, horizontal: 24),
-          decoration: BoxDecoration(
-            color: isDarkMode ? const Color(0xFF6A5AE0).withOpacity(0.2) : const Color(0xFF6A5AE0).withOpacity(0.1),
-            borderRadius: BorderRadius.circular(30),
-          ),
-          child: Text(
-            'You\'re doing great! 👍',
-            style: Theme.of(context).textTheme.titleMedium?.copyWith(
-                  color: isDarkMode ? Colors.amber : const Color(0xFF6A5AE0),
-                  fontWeight: FontWeight.bold,
-                ),
-          ),
-        ),
-      ],
-    );
-  }
+
 }
 
 class _FloatingMathSymbol extends StatefulWidget {
@@ -504,86 +368,6 @@ class _FloatingMathSymbolState extends State<_FloatingMathSymbol>
                 fontSize: 16,
                 fontWeight: FontWeight.bold,
                 color: widget.color,
-              ),
-            ),
-          ),
-        );
-      },
-    );
-  }
-}
-
-class _BouncingMathSymbol extends StatefulWidget {
-  final String symbol;
-  final Color color;
-  final int delay;
-
-  const _BouncingMathSymbol({
-    required this.symbol,
-    required this.color,
-    this.delay = 0,
-  });
-
-  @override
-  State<_BouncingMathSymbol> createState() => _BouncingMathSymbolState();
-}
-
-class _BouncingMathSymbolState extends State<_BouncingMathSymbol>
-    with SingleTickerProviderStateMixin {
-  late AnimationController _controller;
-  late Animation<double> _animation;
-
-  @override
-  void initState() {
-    super.initState();
-    _controller = AnimationController(
-      duration: const Duration(milliseconds: 1500),
-      vsync: this,
-    );
-
-    _animation = TweenSequence<double>([
-      TweenSequenceItem(tween: Tween<double>(begin: 0, end: -15), weight: 50),
-      TweenSequenceItem(tween: Tween<double>(begin: -15, end: 0), weight: 50),
-    ]).animate(
-      CurvedAnimation(
-        parent: _controller,
-        curve: Curves.easeInOut,
-      ),
-    );
-
-    Future.delayed(Duration(milliseconds: widget.delay), () {
-      if (mounted) _controller.repeat(reverse: true);
-    });
-  }
-
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return AnimatedBuilder(
-      animation: _animation,
-      builder: (context, child) {
-        return Transform.translate(
-          offset: Offset(0, _animation.value),
-          child: Container(
-            margin: const EdgeInsets.symmetric(horizontal: 8),
-            child: Text(
-              widget.symbol,
-              style: TextStyle(
-                fontSize: 32,
-                fontWeight: FontWeight.bold,
-                color: widget.color,
-                shadows: [
-                  Shadow(
-                    color: Colors.black.withOpacity(0.1),
-                    blurRadius: 4,
-                    offset: const Offset(2, 2),
-                  ),
-                ],
               ),
             ),
           ),
