+++ lib/main.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:hive_flutter/hive_flutter.dart';
-import 'app/app.dart';
+import 'app/fun_math/app.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
