+++ pubspec.yaml
@@ -79,7 +79,6 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
     - assets/images/
-    - assets/gif/
     - assets/operation/
 
   # An image asset can refer to one or more resolution-specific "variants", see
